Fixed / /= % and %= operators for Int256 and UInt256. Optimized % and %= using AVX2 for Int256.

+++ src/tests/perf/new_perf_tests.cpp
@@ -40,7 +40,7 @@ bool verify_results(const std::vector<int> &results) {
     
     std::ifstream resultFile("int_result.bin", std::ios_base::binary);
     if(resultFile.good()){
-        uint64_t index = 0;
+        unsigned long long index = 0;
         while(index < results.size()) {
 
             resultFile.read((char*)buffer.data(), buffer.size() * sizeof(int));
